requirements.txt
openai
python-dotenv
langchain-nvidia-ai-endpoints
langchain_community
faiss-cpu
streamlit
pypdf
tiktoken